Renaming: DynamicArray → UniqueArray; ArrayModule → BasicContainers
