+++ Package@swift-6.0.swift
@@ -243,7 +243,7 @@ let targets: [CustomTarget] = [
 
   .target(
     kind: .exported,
-    name: "ArrayModule",
+    name: "BasicContainers",
     dependencies: [
       "InternalCollectionsUtilities",
       "ContainersPreview",
@@ -252,9 +252,9 @@ let targets: [CustomTarget] = [
   ),
   .target(
     kind: .test,
-    name: "ArrayTests",
+    name: "BasicContainersTests",
     dependencies: [
-        "ArrayModule", "_CollectionsTestSupport"
+        "BasicContainers", "_CollectionsTestSupport"
     ]),
 
   .target(
