+++ Sources/BasicContainers/RigidArray+Experimental.swift
