+++ Sources/CMakeLists.txt
@@ -58,7 +58,7 @@ endif()
 add_subdirectory(InternalCollectionsUtilities)
 add_subdirectory(ContainersPreview)
 if(NOT COLLECTIONS_FOUNDATION_TOOLCHAIN_MODULE)
-  add_subdirectory(ArrayModule)
+  add_subdirectory(BasicContainers)
   add_subdirectory(BitCollections)
   add_subdirectory(HashTreeCollections)
   add_subdirectory(HeapModule)
